+++ DAL_DataAccessLayer/DAL_Services/DAL_Size.cs
@@ -41,8 +41,13 @@ namespace DAL_DataAccessLayer.DAL_Services
             {
                 using (_db = new QuanLyBanGiayEntities())
                 {
-                    _db.Size.Add(size);
-                    return _db.SaveChanges() > 0 ? "Thêm thành công" : "Thêm thất bại";
+                    if (size != null)
+                    {
+                        _db.Size.Add(size);
+                        _db.SaveChanges();
+                        return "Thêm thành công!";
+                    }
+                    return "Thêm thất bại!";
                 }
             }
             catch (Exception e)
@@ -51,14 +56,26 @@ namespace DAL_DataAccessLayer.DAL_Services
             }
         }
 
-        public string RemoveSize(Size size)
+        public string RemoveSize(string id)
         {
             try
             {
                 using (_db = new QuanLyBanGiayEntities())
                 {
-                    _db.Size.Remove(size);
-                    return _db.SaveChanges() > 0 ? "Xóa thành công" : "Xóa thất bại";
+                    if (id != null)
+                    {
+                        var size = _db.Size.FirstOrDefault(c => c.SizeId == id);
+                        var productDetail = _db.ProductDetail.FirstOrDefault(c => c.SizeId == id);
+                        var inventory = productDetail is null ? null : _db.Inventory.FirstOrDefault(c => c.ProductId == productDetail.ProductId);
+                        var product = productDetail is null ? null : _db.Product.FirstOrDefault(c => c.ProductId == productDetail.ProductId);
+                        if (inventory != null) _db.Inventory.Remove(inventory);
+                        if (productDetail != null) _db.ProductDetail.Remove(productDetail);
+                        if (product != null) _db.Product.Remove(product);
+                        if (size != null) _db.Size.Remove(size);
+                        _db.SaveChanges();
+                        return "Xóa thành công!";
+                    }
+                    return "Xóa thất bại!";
                 }
             }
             catch (Exception e)
@@ -71,16 +88,17 @@ namespace DAL_DataAccessLayer.DAL_Services
         {
             try
             {
-                var result = GetSizeById(size.SizeId);
-                if (result != null)
+                using (_db = new QuanLyBanGiayEntities())
                 {
-                    using (_db = new QuanLyBanGiayEntities())
+                    var result = _db.Size.FirstOrDefault(c => c.SizeId == size.SizeId);
+                    if (size != null && result != null)
                     {
-                        _db.Size.Remove(size);
-                        return _db.SaveChanges() > 0 ? "Sửa thành công" : "Sửa thất bại";
+                        result.SizeName = size.SizeName;
+                        _db.SaveChanges();
+                        return "Sửa thành công!";
                     }
+                    return "Sửa thất bại!";
                 }
-                return "Lỗi không xác định";
             }
             catch (Exception e)
             {
